refactor: 불필요한 useComment 훅 제거
